+++ docs/users/changelog.md
@@ -13,6 +13,7 @@ Note that there is currently no guarantee for a stable Markdown formatting style
     With this plugins can now read CLI arguments merged with values from `.mdformat.toml`.
 - Changed
   - Style: No longer escape square bracket enclosures.
+  - Style: No longer escape less than sign followed by space character.
 - Improved
   - Plugin interface: A trailing newline is added to fenced code blocks if a plugin fails to add it.